pyproject.toml
[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gec-tt-backend"
version = "0.1.0"
description = "Backend for the Tatar GEC proxy"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "starlette>=0.49.1",
    "uvicorn[standard]==0.32.1",
    "python-dotenv==1.0.1",
    "prometheus-client==0.21.0",
    "gunicorn==22.0.0",
]

[tool.setuptools]
packages = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "BLE"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"backend/main.py" = ["B008"]

[tool.mypy]
python_version = "3.11"
files = ["backend"]
check_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[tool.bandit]
exclude_dirs = ["client", "backend/tests"]
